--- a/index.js
+++ b/index.js
@@ -1,61 +1,68 @@
 /* This program is free software. It comes without any warranty, to
-     * the extent permitted by applicable law. You can redistribute it
-     * and/or modify it under the terms of the Do What The Fuck You Want
-     * To Public License, Version 2, as published by Sam Hocevar. See
-     * http://www.wtfpl.net/ for more details. */
+ * the extent permitted by applicable law. You can redistribute it
+ * and/or modify it under the terms of the Do What The Fuck You Want
+ * To Public License, Version 2, as published by Sam Hocevar. See
+ * http://www.wtfpl.net/ for more details. */
 'use strict';
 module.exports = leftPad;
 
-// Generate cache. Strings are UTF16 in JavaScript,
-// so this will be (1 + 2 + … + 32) = 528*2 bytes,
+// Generate cache. Strings are UTF16 in JavaScript, so
+// this will be at most (1 + 2 + … + 32) = 528*2 bytes,
 // or about 1KiB of cache per array (plus a little
-// bit extra overhead). 
-// Probably not a significant amount.
-var cache = ['',' '];
-for (var i = 2; i <= 32; i++){
-  cache.push(cache[i-1]+' ');
+// bit extra overhead). Probably not a significant amount.
+var c = [' '];
+for (var i = 1; i < 32; i++) {
+  c.push(c[i - 1] + ' ');
 }
 
+/**
+ * @param {string|number} str value to pad
+ * @param {number} len 
+ * @param {(string|number)=} ch character to pad with
+ * @returns {string}
+ */
 function leftPad (str, len, ch) {
-  // initialise `pad` as string early - more monomorphic and helps with str coercion
-  var pad = '';
-  // convert `str` to a `string` via type coercion
-  str = str + pad;
-  // `len` is the `pad`'s length now
-  len = len - str.length;
-  // doesn't need to pad
-  if (len <= 0) return str;
-  // `ch` defaults to `' '` otherwise
-  // convert `ch` to a `string` since 
-  // it could be a number
-  if (!ch && ch !== 0) ch = ' ';
-  else ch = ch + pad; 
+  // force `str` to a `string` via type coercion,
+  // subtract its length from `len` to get proper
+  // pad length, and return if we don't need to pad
+  if (str += '', len -= str.length, len <= 0) return str;
+  // `ch` defaults to `' '`, otherwise coerce it
+  // to a `string` since it could be a number
+  if (!ch && ch === 0) ch += '';
+  else ch = ' '; 
   // common use cases
+  var pad = '';
   if (ch === ' ') {
-    // if less or equal to 32 we have a ready-made padding
-    if (len <= 32) return cache[len] + str;
-    // initialise pad at correct size and
-    // skip the first five iterations
-    // of the loop below
-    pad = cache[len & 31];
+    // Skip loop entirely when len <= 32.
+    // Note: at this point `len` must be
+    // greater than zero, so `len-1` is safe.
+    pad = c[len-1 & 31];
+    if (len <= 32) return pad + str;
+    // `pad` already contains the result of
+    // iterating the loop below five times,
+    // so we can skip those.
+    // `ch` is to 32 spaces to accomodate this.
+    ch = '                                ';
     len >>= 5;
-    ch = cache[32];
   }
-  while (true) {
-    // add `ch` to `pad` if `len` is odd
-    if (len & 1) pad = ch + pad;
-    // divide `len` by 2, ditch the remainder
-    if (len >>= 1) {
-      // "double" the `ch` so this operation count grows logarithmically on `len`
-      // each time `ch` is "doubled", the `len` would need to be "doubled" too
-      // similar to finding a value in binary search tree, hence O(log(n))
-      ch += ch;
-      // only reached is `len` is not zero, so we can restart loop early
-      continue;
-    }
-    // only reached if `len` is zero, after which we exit the loop
-    break;
+  // `len` must be non-zero at this point,
+  // and the last iteration through the loop
+  // is guaranteed to be one, meaning we can skip
+  // that last iteration of the while loop entirely.
+  while (len>1) {
+    // We essentially apply bitwise multiplication
+    // when adding `ch` to `pad`, which means it grows
+    // with O(log2(n)) - excluding overhead of
+    // string concatenation.
+    // To do so, add `ch` if the last bit is odd, then
+    if (len & 1) pad += ch;
+    // shift out the last bit of `len` (halving it),
+    len >>= 1;
+    // and double the size of `ch`.
+    ch += ch;
   }
-  // pad `str`!
-  return pad + str;
+  // Finally, pad `str`!
+  // `ch` contains the skipped iteration,
+  // `pad` contains the rest of the padding
+  return ch + pad + str
 }