--- a/index.js
+++ b/index.js
@@ -6,13 +6,14 @@
 'use strict';
 module.exports = leftPad;
 
-// Generate cache. Strings are UTF16 in JavaScript, so
-// this will be at most (1 + 2 + … + 32) = 528*2 bytes,
-// or about 1KiB of cache per array (plus a little
-// bit extra overhead). Probably not a significant amount.
-var c = [' '];
+// Generate cache for common use-case of padding with spaces and zeros.
+// JavaScript has UTF16 strings, so worst case this is 2 bytes times
+// 1 + 2 + … + 32 = 528 chars, or about 1KiB per array, plus a few
+// bytes of overhead per string. Probably not a significant amount.
+var c_ = [' '], c0 = ['0'];
 for (var i = 1; i < 32; i++) {
-  c.push(c[i - 1] + ' ');
+  c_.push(c_[i - 1] + ' ');
+  c0.push(c0[i - 1] + '0');
 }
 
 /**
@@ -22,30 +23,32 @@ for (var i = 1; i < 32; i++) {
  * @returns {string}
  */
 function leftPad (str, len, ch) {
-  // force `str` to a `string` via type coercion,
-  // subtract its length from `len` to get proper
+  // Force `str` to a string via type coercion,
+  // subtract `str.length` from `len` to get proper
   // pad length, and return if we don't need to pad
   if (str += '', len -= str.length, len <= 0) return str;
-  // `ch` defaults to `' '`, otherwise coerce it
-  // to a `string` since it could be a number
-  if (!ch && ch !== 0) ch = ' ';
-  else ch += ''; 
-  // common use cases
   var pad = '';
-  if (ch === ' ') {
-    pad = c[len-1 & 31];
+  // `ch` defaults to `' '` for falsy values, except for 0
+  if (ch === ' ' || !ch && ch !== 0){
+    // We use a cache to speed up the default case
+    pad = c_[len-1 & 31];
     if (len <= 32) return pad + str;
-    // `pad` already contains the result of
+    // `pad` effectively contains the result of
     // iterating the loop below five times,
     // so we can skip those.
-    // `ch` is to 32 spaces to accomodate this.
+    // `ch` is set to 32 spaces to accomodate this.
     ch = '                                ';
     len >>= 5;
+  } else if ((ch+= '') === '0'){
+    // padding by zero is another really common use-case
+    pad = c0[len-1 & 31];
+    if (len <= 32) return pad + str;
+    ch = '00000000000000000000000000000000';
+    len >>= 5;
   }
-  // `len` must be non-zero at this point,
-  // and the last iteration through the loop
-  // is guaranteed to be one, meaning we can skip
-  // that last iteration of the while loop entirely.
+  // `len` must be non-zero at this point, and the
+  // last loop iteration is guaranteed to be one,
+  // meaning we can skip testing for it entirely.
   while (len>1) {
     // We essentially apply bitwise multiplication
     // when adding `ch` to `pad`, which means it grows