So I just spent entirely too much time this evening/night trying to make `left-pad` faster, as can be seen by [the number of benchmarks I've created on perf.zone](https://perf.zone/user/tlh545IWtPgS86I7SAF8Qjma6w32):
![leftpad-2](https://user-images.githubusercontent.com/259840/36764219-dac48b8a-1c2b-11e8-85f3-bba14d21e2f3.gif)

After various dead ends (alternate ways of breaking out of the loop, loop unrolling), I finally found a set of tweaks that leads to consistent improvements for longer padding, _without_ reduction in performance when padding only a handful of characters.

There are a few small tweaks that had consistent improvements, like hoisting the padding variable, and switching the order in which `ch` is added to `pad`, because [it turns out](https://run.perf.zone/view/String-concat-does-the-order-matter-1519755302670) to matter [a tiny bit](https://run.perf.zone/view/String-concat-does-the-order-matter-1-190-1519755398884). Furthermore, instead of *breaking out* of the while-loop when `len` is zero, we *continue* it early as long as `len` is greater than zero.

However, the biggest change is very clearly this:

```js
    // ==== OLD ====
    if (ch === ' ' && len < 10) return cache[len] + str;
    var pad = '';
   
    // ==== NEW ====
    if (ch === ' ') {
        if (len < 16) return cache[len] + str;
        pad = cache[len & 15];
        len >>= 4;
        ch = '                ';
    }
```

Expanding the cache let's use skip the while loop when padding 10-15 characters. That's just a side-effect of the code below it, which leads to enormous improvements when padding large numbers of characters: `pad = cache[len & 15];  len >>= 4; ch = '                ';`
This initiates `pad` at the right size to let us skip the first four iterations of the loop. Recall that padding by 256 characters adds up to eight iterations. So this halved the time it takes to do that (minus a little bit of overhead).

Note that this would also work with `pad = cache[len & 7];  len >>= 3; ch = '        ';`, and in fact the benchmarks include that version. As you can see below, the 16-element version gives a much, _much_ higher benefit, and IMO isn't so much oversized as the next step (32 elements) would be.

# Selected, probably-bug-free benchmarks

## neither regression nor improvement
- [pad none](https://run.perf.zone/view/pad-none-leftpad-bugfix-Steve-Maos-baseline-more-variants-1519783877604), 
- [pad 4](https://run.perf.zone/view/pad-4-leftpad-bugfix-Steve-Maos-baseline-more-variants-1519781681326) (run multiple times to see how the "winner" varies each run), 
- [pad 8](https://run.perf.zone/view/pad-8-leftpad-bugfix-Steve-Maos-baseline-more-variants-1519781475410)

## Significant perf boosts

- [pad 12](https://run.perf.zone/view/pad-12-leftpad-bugfix-Steve-Maos-baseline-more-variants-1519781278596) (kind of"cheating", since we bumped the cache size)
- [pad 25](https://run.perf.zone/view/pad-25-leftpad-bugfix-Steve-Maos-baseline-more-variants-1519781088832)
- [pad 50](https://run.perf.zone/view/pad-50-leftpad-bugfix-Steve-Maos-baseline-more-variants-1519780882720)
- [pad 90](https://run.perf.zone/view/pad-90-leftpad-bugfix-Steve-Maos-baseline-more-variants-1519780543048)
- [pad 9001](https://run.perf.zone/view/pad-9001-leftpad-bugfix-Steve-Maos-baseline-more-variants-1519780208772)

# Benchmark measurement gallery

## Chrome (Windows 10 64 bits)
![image](https://user-images.githubusercontent.com/259840/36765268-a4bca89c-1c30-11e8-8f93-f8f91b8e3661.png)

![image](https://user-images.githubusercontent.com/259840/36765275-b0951302-1c30-11e8-8904-e786aa8c0976.png)

![image](https://user-images.githubusercontent.com/259840/36765279-b6a435f2-1c30-11e8-9fb1-73135be472fe.png)

![image](https://user-images.githubusercontent.com/259840/36765283-bcc070c2-1c30-11e8-837e-40150468928d.png)

![image](https://user-images.githubusercontent.com/259840/36766080-8e527768-1c34-11e8-9bd6-10ee986335b7.png)

## Firefox 58 (Windows 10, 64 bits)

![image](https://user-images.githubusercontent.com/259840/36765349-05266a6a-1c31-11e8-8021-25c463ef4bbc.png)

![image](https://user-images.githubusercontent.com/259840/36765343-fd5896dc-1c30-11e8-84df-27cd7b63cb0a.png)

![image](https://user-images.githubusercontent.com/259840/36765333-f4eca0b0-1c30-11e8-87a3-15d529e10754.png)

![image](https://user-images.githubusercontent.com/259840/36765326-ec5db3da-1c30-11e8-9b0e-ca72ed24c2e7.png)

![image](https://user-images.githubusercontent.com/259840/36765899-c735badc-1c33-11e8-8376-d87350649cb9.png)