@JobLeonard 2018-02-28
Made cache generated - it's less characters, and caching up to length 32 adds up to about a kilobyte of memoised strings. Probably not a significant overhead, but a hu gain in perf in many circumstances.

The following benchmark also include a 256-sized cache, which would be around 64 KiB of memoisation overhead. Given that it's not a very common use-case, I stuck to the maximum length of 32.

[pad 20](https://run.perf.zone/view/pad-20-leftpad-humongous-cache-1519791225526)
[pad 10](https://run.perf.zone/view/pad-10-leftpad-humongous-cache-1519791398388)
[pad 5](https://run.perf.zone/view/pad-5leftpad-humongous-cache-1519791633330)
[pad 500](https://run.perf.zone/view/pad-500-leftpad-humongous-cache-1519791827316)