@zehua 2018-02-28
can you also try the perf tests in the perf folder? and test in node.js?