@benjamingr 2018-02-28
So, I don't really trust these benchmarks :D 

Would you be able to run them using the Node benchmarking tool so we can see statistical significance? Alternatively, it would be nice to see an asm dump from v8.