@benjamingr 2018-02-28
That said, to clarify, I believe this is a speed improvement - just not digging the benchmarks :D