@JobLeonard 2018-03-01
@zehua: I had a suspicion that the bit patterns may have an influence on performance, so I even added a bunch of extra tests for that:

```
10922 (0b10101010101010)
 8193 (0b10000000000001)
 8192 (0b10000000000000)
 8191  (0b1111111111111)
 5461  (0b1010101010101)
 4096  (0b1000000000000)
 4095   (0b111111111111)
 2079   (0b100000011111)
 2048   (0b100000000000)
 2047    (0b11111111111)
 1365    (0b10101010101)
 1025    (0b10000000001)
 1024    (0b10000000000)
 1023     (0b1111111111)
```

It turns out the bit patterns do match: the more bits are set, the slower the algorithms are (except for `O(n)`). Makes sense, as these cases require more string concatenation after all.

I also tried a whole bunch of alternative tweaks, but none of them worked out:

```js
  // New while-loop
  while (true) {
    if (len & 1) pad = ch + pad;
    if (len >>= 1) {
      ch += ch;
      continue;
    }
    break;
  }

  // Switch, two bits. Three bits is similar, but with cases 7 to 0
  while (true) {
    ch2 = ch + ch;
    switch(len&3){
      case 3: //
        pad = ch2 + ch + pad;
        break;
      case 2:
        pad = ch2 + pad;
        break;
      case 1:
        pad = ch + pad;
      case 0:
      default:
    }
    if (len >>= 2) {
      ch = ch2 + ch2;
      continue;
    }
    break;
  }

  // Array lookup. The idea being that maybe branching was the bottleneck
  // var c = ['', '', '', ''];
  while (true) {
    c[2] = c[1] + c[1];
    pad = c[len&2] + (c[len&1] + pad);
    if (len >>= 2) {
      c[1] = c[2] + c[2];
      continue;
    }
    break;
  }
``` 


Anyway, the output of running all of the performance tests, after filtering the discarded alternative implementations:

```
-> pad 100 spaces to str of len 4
O(n)         x 1,578,867 ops/sec ±1.27% (90 runs sampled)
ES6 Repeat   x 8,546,499 ops/sec ±0.86% (89 runs sampled)
Current      x 8,829,279 ops/sec ±0.85% (90 runs sampled)
New          x 12,440,392 ops/sec ±0.57% (92 runs sampled)
-> pad  33 spaces to str of len 4
O(n)         x 3,493,611 ops/sec ±0.51% (92 runs sampled)
ES6 Repeat   x 8,907,255 ops/sec ±0.62% (95 runs sampled)
Current      x 9,583,444 ops/sec ±0.45% (94 runs sampled)
New          x 19,276,613 ops/sec ±0.63% (93 runs sampled)
-> pad  32 spaces to str of len 4
O(n)         x 3,549,147 ops/sec ±0.64% (90 runs sampled)
ES6 Repeat   x 9,081,341 ops/sec ±0.66% (93 runs sampled)
Current      x 9,712,771 ops/sec ±0.57% (93 runs sampled)
New          x 27,114,519 ops/sec ±0.62% (92 runs sampled)
-> pad  31 spaces to str of len 4
O(n)         x 3,536,994 ops/sec ±1.06% (89 runs sampled)
ES6 Repeat   x 7,672,999 ops/sec ±0.59% (89 runs sampled)
Current      x 8,131,713 ops/sec ±0.81% (92 runs sampled)
New          x 25,577,231 ops/sec ±1.56% (88 runs sampled)
-> pad  10 spaces to str of len 4
O(n)         x 5,901,201 ops/sec ±0.74% (95 runs sampled)
ES6 Repeat   x 9,047,213 ops/sec ±0.53% (95 runs sampled)
Current      x 9,625,336 ops/sec ±0.72% (92 runs sampled)
New          x 24,329,822 ops/sec ±0.89% (87 runs sampled)
-> pad   9 spaces to str of len 4
O(n)         x 6,201,515 ops/sec ±0.52% (94 runs sampled)
ES6 Repeat   x 8,912,250 ops/sec ±0.95% (93 runs sampled)
Current      x 22,530,869 ops/sec ±1.02% (89 runs sampled)
New          x 23,041,314 ops/sec ±0.77% (91 runs sampled)
-> pad  no spaces to str of len 4
O(n)         x 29,504,068 ops/sec ±0.88% (88 runs sampled)
ES6 Repeat   x 30,030,613 ops/sec ±1.15% (90 runs sampled)
Current      x 30,261,165 ops/sec ±0.96% (86 runs sampled)
New          x 29,898,400 ops/sec ±1.00% (90 runs sampled)


-> pad 100 spaces to str of len 100
O(n)         x 1,727,331 ops/sec ±0.75% (95 runs sampled)
ES6 Repeat   x 8,180,206 ops/sec ±0.81% (93 runs sampled)
Current      x 8,596,671 ops/sec ±0.59% (94 runs sampled)
New          x 11,964,140 ops/sec ±0.73% (92 runs sampled)
-> pad  10 spaces to str of len 100
O(n)         x 9,257,534 ops/sec ±0.60% (92 runs sampled)
ES6 Repeat   x 8,906,042 ops/sec ±0.75% (93 runs sampled)
Current      x 9,475,165 ops/sec ±0.67% (90 runs sampled)
New          x 21,960,796 ops/sec ±0.82% (87 runs sampled)
-> pad   9 spaces to str of len 100
O(n)         x 9,579,826 ops/sec ±0.88% (92 runs sampled)
ES6 Repeat   x 8,868,416 ops/sec ±0.65% (92 runs sampled)
Current      x 23,086,294 ops/sec ±0.91% (91 runs sampled)
New          x 22,893,180 ops/sec ±0.78% (92 runs sampled)
-> pad   0 spaces to str of len 100
O(n)         x 15,015,491 ops/sec ±0.91% (90 runs sampled)
ES6 Repeat   x 12,319,054 ops/sec ±0.68% (86 runs sampled)
Current      x 23,131,907 ops/sec ±0.89% (90 runs sampled)
New          x 22,928,214 ops/sec ±0.75% (88 runs sampled)


-> pad 10922 (0b10101010101010) spaces to str of len 4
O(n)         x 141,780 ops/sec ±0.94% (92 runs sampled)
ES6 Repeat   x 6,245,441 ops/sec ±0.71% (88 runs sampled)
Current      x 6,520,407 ops/sec ±0.59% (93 runs sampled)
New          x 7,412,821 ops/sec ±0.73% (92 runs sampled)
-> pad  8193 (0b10000000000001) spaces to str of len 4
O(n)         x 20,789 ops/sec ±1.04% (88 runs sampled)
ES6 Repeat   x 6,503,653 ops/sec ±0.71% (92 runs sampled)
Current      x 6,663,526 ops/sec ±0.61% (90 runs sampled)
New          x 7,079,254 ops/sec ±0.70% (92 runs sampled)
-> pad  8192 (0b10000000000000) spaces to str of len 4
O(n)         x 20,953 ops/sec ±1.03% (89 runs sampled)
ES6 Repeat   x 6,718,228 ops/sec ±0.70% (91 runs sampled)
Current      x 6,955,801 ops/sec ±0.82% (92 runs sampled)
New          x 7,116,594 ops/sec ±0.68% (92 runs sampled)
-> pad  8191  (0b1111111111111) spaces to str of len 4
O(n)         x 20,928 ops/sec ±0.97% (91 runs sampled)
ES6 Repeat   x 4,664,857 ops/sec ±0.94% (94 runs sampled)
Current      x 4,834,961 ops/sec ±0.63% (93 runs sampled)
New          x 5,155,518 ops/sec ±0.65% (91 runs sampled)


-> pad  5461  (0b1010101010101) spaces to str of len 4
O(n)         x 31,568 ops/sec ±0.87% (86 runs sampled)
ES6 Repeat   x 5,741,766 ops/sec ±0.80% (89 runs sampled)
Current      x 5,977,551 ops/sec ±0.54% (93 runs sampled)
New          x 6,505,239 ops/sec ±0.68% (92 runs sampled)
-> pad  4096  (0b1000000000000) spaces to str of len 4
O(n)         x 42,545 ops/sec ±1.03% (90 runs sampled)
ES6 Repeat   x 6,945,246 ops/sec ±0.60% (90 runs sampled)
Current      x 7,264,366 ops/sec ±0.82% (94 runs sampled)
New          x 7,616,480 ops/sec ±0.47% (93 runs sampled)
-> pad  4095   (0b111111111111) spaces to str of len 4
O(n)         x 42,363 ops/sec ±1.06% (90 runs sampled)
ES6 Repeat   x 5,191,652 ops/sec ±0.82% (95 runs sampled)
Current      x 5,241,190 ops/sec ±1.54% (85 runs sampled)
New          x 5,544,186 ops/sec ±1.22% (93 runs sampled)


-> pad  2079   (0b100000011111) spaces to str of len 4
O(n)         x 92,721 ops/sec ±0.79% (90 runs sampled)
ES6 Repeat   x 5,952,327 ops/sec ±0.76% (87 runs sampled)
Current      x 6,189,434 ops/sec ±0.51% (94 runs sampled)
New          x 7,888,843 ops/sec ±0.71% (95 runs sampled)
-> pad  2048   (0b100000000000) spaces to str of len 4
O(n)         x 94,377 ops/sec ±0.67% (94 runs sampled)
ES6 Repeat   x 7,143,774 ops/sec ±0.57% (91 runs sampled)
Current      x 7,405,973 ops/sec ±0.79% (93 runs sampled)
New          x 8,260,003 ops/sec ±0.56% (93 runs sampled)
-> pad  2047    (0b11111111111) spaces to str of len 4
O(n)         x 93,449 ops/sec ±0.83% (93 runs sampled)
ES6 Repeat   x 5,468,892 ops/sec ±0.55% (94 runs sampled)
Current      x 5,630,172 ops/sec ±0.75% (88 runs sampled)
New          x 6,671,885 ops/sec ±0.57% (94 runs sampled)


-> pad  1365    (0b10101010101) spaces to str of len 4
O(n)         x 142,199 ops/sec ±0.71% (90 runs sampled)
ES6 Repeat   x 6,251,298 ops/sec ±0.75% (92 runs sampled)
Current      x 6,549,310 ops/sec ±0.52% (94 runs sampled)
New          x 7,673,830 ops/sec ±0.74% (93 runs sampled)
-> pad  1025    (0b10000000001) spaces to str of len 4
O(n)         x 188,464 ops/sec ±0.69% (95 runs sampled)
ES6 Repeat   x 7,169,215 ops/sec ±0.61% (89 runs sampled)
Current      x 7,510,027 ops/sec ±0.65% (95 runs sampled)
New          x 8,596,675 ops/sec ±0.62% (91 runs sampled)
-> pad  1024    (0b10000000000) spaces to str of len 4
O(n)         x 178,073 ops/sec ±1.60% (87 runs sampled)
ES6 Repeat   x 7,082,301 ops/sec ±1.42% (87 runs sampled)
Current      x 7,545,679 ops/sec ±2.29% (87 runs sampled)
New          x 8,483,300 ops/sec ±1.18% (87 runs sampled)
-> pad  1023     (0b1111111111) spaces to str of len 4
O(n)         x 189,300 ops/sec ±0.74% (91 runs sampled)
ES6 Repeat   x 5,775,324 ops/sec ±0.69% (91 runs sampled)
Current      x 5,836,955 ops/sec ±1.89% (90 runs sampled)
New          x 7,430,372 ops/sec ±0.77% (92 runs sampled)
```

[perf-2018-03-01.txt](https://github.com/stevemao/left-pad/files/1769563/perf-2018-03-01.txt)